// app/api/protocol/gql/graphql.ts
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  BigDecimal: string;
  BigInt: bigint;
  Bytes: string;
  /**
   * 8 bytes signed integer
   *
   */
  Int8: number;
  /**
   * A string representation of microseconds UNIX timestamp (16 digits)
   *
   */
  Timestamp: string;
};
// ...
export type Price = {
  __typename?: 'Price';
  /**
   * The decimals of the "value".
   *
   */
  decimals: Scalars['Int'];
  /**
   * The ID of the price (same as Token ID).
   *
   */
  id: Scalars['Bytes'];
  /**
   * The actual price of the token (scaled up to "decimals").
   *
   */
  value: Scalars['BigInt'];
};
// ...
export type Token = {
  __typename?: 'Token';
  /**
   * The blockchain address of the asset.
   *
   */
  address: Scalars['String'];
  /**
   * The number of decimal places for the asset.
   *
   */
  decimals: Scalars['Int'];
  /**
   * The ID of the token. Mainly used for subgraph internal indexing.
   *
   */
  id: Scalars['Bytes'];
  /**
   * The name of the asset.
   *
   */
  name: Scalars['String'];
  /**
   * The price of the token.
   *
   */
  price: Price;
  /**
   * The symbol of the asset.
   *
   */
  symbol: Scalars['String'];
};